# app/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Management System",
    "description": "Create and Manage Users, Roles and Permissions",
    "contact": {
      "name": "Bibek Aryal"
    },
    "license": {
      "name": "Personal Use Only"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/ums",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/apps/app/{id}": {
      "get": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Get app by ID",
        "description": "Retrieves a specific app by its unique ID",
        "operationId": "readApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Update an app by ID",
        "description": "Updates an existing app identified by its unique ID",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Details of the app to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "App updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Soft delete an app by ID",
        "description": "Marks an app as deleted without permanently removing it, identified by its unique ID",
        "operationId": "softDeleteApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to soft delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_users/user/{id}": {
      "get": {
        "tags": [
          "Users Management"
        ],
        "summary": "Retrieve app user by user ID",
        "description": "Fetches details of a specific app user by user ID.",
        "operationId": "readAppUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the specified app user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Not Authorized or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User Not Found by specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users Management"
        ],
        "summary": "Update app user details",
        "description": "Updates details of a specific app user by ID.",
        "operationId": "updateAppUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app user to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated details of the app user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated details of the app user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Not Authorized or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User to update Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users Management"
        ],
        "summary": "Soft delete an app user",
        "description": "Marks a permission as deleted without permanently removing it, identified by its unique ID",
        "operationId": "softDeleteAppUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app user to soft delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Not Authorized or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User to soft delete Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps_app_user/apps_user": {
      "post": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Assign User to an App",
        "description": "Adds a User to an App",
        "operationId": "createAppsAppUser",
        "requestBody": {
          "description": "Details of user and app to be linked",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppsAppUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User assigned successfully to app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User or App to Link Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps_app_user": {
      "get": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Get all apps and users linked",
        "description": "Retrieves all linked apps and users available in the system",
        "operationId": "readAppsAppUsers",
        "responses": {
          "200": {
            "description": "Apps Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/apps_app_user/user/{appUserId}": {
      "get": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Get linked apps and users by user ID",
        "description": "Retrieves all apps and users associated with the specified user ID",
        "operationId": "readAppsAppUsersByUserId",
        "parameters": [
          {
            "name": "appUserId",
            "in": "path",
            "description": "ID of the user for which apps users are retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apps Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/apps_app_user/app/{appId}": {
      "get": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Get linked apps and users by app ID",
        "description": "Retrieves all apps and users associated with the specified app ID",
        "operationId": "readAppsAppUsersByAppId",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "description": "ID of the app for which apps users are retrieved",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apps Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/apps_app_user/app/{appId}/user/{appUserEmail}": {
      "get": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Get a linked app user by app ID and user email",
        "description": "Retrieves a specific app user its app ID and user email",
        "operationId": "readAppsAppUsersByAppIdAndUserEmail",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "description": "ID of the app in linked app user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appUserEmail",
            "in": "path",
            "description": "Email of the User in linked app user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App User Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps_app_user/apps_user/{appId}/{appUserEmail}": {
      "delete": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Unassign User from an App",
        "description": "Removes a User from an App",
        "operationId": "deleteAppsAppUser",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "description": "ID of the Application in linked app user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appUserEmail",
            "in": "path",
            "description": "Email of the User in linked app user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User unassigned successfully from app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppsRequest": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "AppsResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppsDto"
            }
          }
        }
      },
...
      "ResponseStatusInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errMsg": {
            "type": "string"
          }
        }
      },
...
      "AppUserRequest": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "status"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "guestUser": {
            "type": "boolean"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppUserAddressDto"
            }
          }
        }
      },
...
      "AppUserResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppUserDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Token": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "Basic": {
        "type": "http",
        "scheme": "Basic"
      }
    }
  }
}